requirements/production.txt
django
requests
Pillow
hashids
python-social-auth
django-pipeline
djangorestframework
celery
redis
flower
